+++ lib/appsignal/transaction/params_sanitizer.rb
@@ -53,6 +53,9 @@ module Appsignal
 
         def inspected(value)
           value.inspect
+        rescue
+          # It turns out that sometimes inspect can fail
+          "#<#{value.class.to_s}/>"
         end
       end
     end